Экран мема

+++ i-memes/src/main/java/com/example/i_memes/response/MemesResponse.kt
@@ -6,7 +6,7 @@ import ru.surfstudio.standard.i_network.network.Transformable
 
 data class MemesResponse(
         @SerializedName("id") val id: String,
-        @SerializedName("createdDate") val createdDate: Int,
+        @SerializedName("createdDate") val createdDate: Long,
         @SerializedName("description") val description: String,
         @SerializedName("isFavorite") val isFavorite: Boolean,
         @SerializedName("photoUtl") val photoUtl: String,
